Update tests related to looping empty tuple
